// src/app/models/pokemon.ts
export class Pokemon {
  name: string;
  id: number;
  sprites: {};
  types: {} [];
  type: string;
  type2: string;
  weight: string;
  height: string;
  img: string;
  imgShiny: string;
  typeId: number;
  stats: {} [];
  hp: string;
  attack: string;
  defense: string;
  specialAttack: string;
  specialDefense: string;
  speed: string;
  back: string;
  poke: { types: { slot: string; type: { name: string; url: string; }; }[]; };

}